nl fix

+++ libs/yocto_extension/yocto_extension.h
@@ -54,7 +54,7 @@ namespace yocto::extension {
     
     void denoise(float* colorPtr, float* outputPtr, int width, int heigh);
 
-    img::image<vec4f> denoise_nlmean(const img::image<vec4f>& img, int r, int f, int sigma, int h);
+    img::image<vec4f> denoise_nlmean(const img::image<vec4f>& img, int r, int f, float sigma, float h);
 
 }
